requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
chromadb>=0.4.22
openai>=1.6.1
tiktoken>=0.5.2
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchainhub>=0.1.14
langchain-nomic>=0.0.3
nomic[local]>=3.0.11
langchain-ollama>=0.0.5
scikit-learn>=1.3.2
tavily-python>=0.3.0 